+++ project/src/components/login-page/login-form.tsx
@@ -1,0 +1,56 @@
+import { SyntheticEvent, useState, memo } from 'react';
+import { useAppDispatch } from '../../hooks/useAppDispatch';
+import { loginAction } from '../../store/api-action';
+import { toast } from 'react-toastify';
+
+function LoginForm(): JSX.Element {
+  const dispatch = useAppDispatch();
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+
+  const handleSubmit = (e: SyntheticEvent) => {
+    e.preventDefault();
+    if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email)) {
+      toast.error('Email is invalid');
+      return;
+    }
+    if (!/[a-zA-Z]/.test(password) || !/\d/.test(password) || password.includes(' ')) {
+      toast.error('Password is invalid');
+      return;
+    }
+
+    dispatch(loginAction({ email, password }));
+  };
+
+  return (
+    <form className="login__form form" onSubmit={handleSubmit}>
+      <div className="login__input-wrapper form__input-wrapper">
+        <label className="visually-hidden">E-mail</label>
+        <input
+          className="login__input form__input"
+          type="email"
+          name="email"
+          placeholder="Email"
+          required
+          value={email}
+          onChange={(e) => setEmail(e.target.value)}
+        />
+      </div>
+      <div className="login__input-wrapper form__input-wrapper">
+        <label className="visually-hidden">Password</label>
+        <input
+          className="login__input form__input"
+          type="password"
+          name="password"
+          placeholder="Password"
+          required
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+        />
+      </div>
+      <button className="login__submit form__submit button" type="submit">Sign in</button>
+    </form>
+  );
+}
+
+export default memo(LoginForm);
